+++ doc/scripting/variables.qml
@@ -13,6 +13,10 @@ QtObject {
     
     // register your settings variables so the user can set them in the script settings
     // use this property if you don't need 
+    //
+    // unfortunately there is no QVariantHash in Qt, we only can use
+    // QVariantMap (that has no arbitrary ordering) or QVariantList (which at
+    // least can be ordered arbitrarily)
     property variant settingsVariables: [
         {
             "identifier": "myString",
@@ -47,6 +51,16 @@ QtObject {
     
     function init() {
         script.registerCustomAction("variablesTest", "Variables test", "Var test", "edit-copy");
+        doSomethingWithYourVariables();
+    }
+    
+    function doSomethingWithYourVariables() {
+        // these variables will be set by QOwnNotes from the settings
+        // if the user didn't set a variable your default value will be set
+        script.log(myString);
+        script.log(myText);
+        script.log(myInt);
+        script.log(myFile);
     }
     
     /**
@@ -60,60 +74,19 @@ QtObject {
             return;
         }
 
-        // these variables will be set by QOwnNotes from the settings
-        script.log(myString);
-        script.log(myText);
-        script.log(myInt);
-        script.log(myFile);
-        
-        script.log(settingsVariables[0].identifier);
+        doSomethingWithYourVariables();
     }
 
     /**
-     * Registeres the settings variables
+     * Registers the settings variables again
      * 
-     * This is the 2nd way to register settings variables.
-     * Use this method if you want to use code to set your variables, like setting
+     * Use this method if you want to use code to override your variables, like setting
      * default values dependend on the operating system.
-     * 
-     * registerSettingsVariables will override the settingsVariables property
-     * 
-     * @returns a map with the information about the variables that should be registered
-     *          to be shown in the script settings
      */
     function registerSettingsVariables() {
-        // don't forget to define these variables also as property
-        var variables = [
-            {
-                "identifier": "myString",
-                "name": "I am a line edit",
-                "description": "Please enter a valid string:",
-                "type": "string",
-                "default": "My default value",
-            },
-            {
-                "identifier": "myText",
-                "name": "I am textbox",
-                "description": "Please enter your text:",
-                "type": "text",
-                "default": "This can be a really long text\nwith multiple lines.",
-            },
-            {
-                "identifier": "myInt",
-                "name": "I am a number selector",
-                "description": "Please enter a number:",
-                "type": "integer",
-                "default": 42,
-            },
-            {
-                "identifier": "myFile",
-                "name": "I am a file selector",
-                "description": "Please select the file:",
-                "type": "file",
-                "default": "pandoc",
-            }
-        ];
-        
-        return variables;
+        if (script.platformIsWindows()) {
+            // override the myFile default value
+            settingsVariables[3].default = "pandoc.exe"
+        }
     }
 }